Fix homebrew not installing packages on first run
Current Homebrew script fails to install packages when there's nothing already installed. Manual fix was to install one package manually and then re-run Puppet.
Cause was the return of the `brew cask --versions` command which is never empty since includes warnings and error messages that made the parser fail to work on them due it's string structure.
Added a easy fix to check whether the warming exist, if it does, the return value is set to empty. Tested on brand new MacOSX install running El Capitan.